+++ test-suite/run_tests.ps1
@@ -40,6 +40,7 @@ $ALL_TEST_NAMES = @(
     "TestUnhandledException"
     "TestStdIO"
     "TestModules"
+    "TestTracePoint"
 )
 
 $TEST_NAMES = $tests