+++ test-suite/run_tests.sh
@@ -34,6 +34,7 @@ ALL_TEST_NAMES=(
     "TestUnhandledException"
     "TestStdIO"
     "TestModules"
+    "TestTracePoint"
 )
 
 TEST_NAMES="$@"